refactor: Restructure parsers to extend a common base class
Various other code changes to simplify implementation and allow the parsed list of objects to be returned from the parsers. This allows them to be printed or used in unit tests.

Also modify the command-line parameters to assume first is GPSTest file and second is Chicago TNC CSV file.
