--- src/main/java/edu/usf/cutr/grha/constants/FileNameConstants.java
@@ -1,9 +1,0 @@
-package edu.usf.cutr.grha.constants;
-
-public class FileNameConstants {
-
-    public static final String CHICAGO_DATA_FILE_NAME = "Transportation_Network_Providers_reduced_records.csv";
-
-    public static final String GPS_TEST_DATA_FILE_NAME = "gnss_log_2021_02_05_13_20_58_beans.txt";
-
-}
